refactoring and type/lint issues fix
