+++ packages/web/app/src/pages/organization.tsx
@@ -1,44 +1,23 @@
-import {
-  ChangeEvent,
-  ReactElement,
-  useCallback,
-  useEffect,
-  useM,
-  useMemo,
-  useRef,
-  useState,
-  useStateemo,
-} from 'react';
-import { differenceInMilliseconds, endOfDay, formatISO, startOfDay } from 'date-fns';
-import * as echarts from 'echarts';
-import ReactECharts from 'echarts-for-react';
-import { ArrowUpRight, Globe, History, MoveDownIcon, MoveUpIcon, SearchIcon } from 'lucide-react';
-import AutoSizer from 'react-virtualized-auto-sizer';
+import { ChangeEvent, useCallback, useMemo, useRef } from 'react';
+import { endOfDay, formatISO, startOfDay } from 'date-fns';
+import { ArrowUpRight, MoveDownIcon, MoveUpIcon, SearchIcon } from 'lucide-react';
 import { useQuery } from 'urql';
 import { z } from 'zod';
 import { OrganizationLayout, Page } from '@/components/layouts/organization';
+import { TargetCard, TargetCardSkeleton } from '@/components/organization/TargetCard';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import { Card } from '@/components/ui/card';
 import { EmptyList } from '@/components/ui/empty-list';
 import { Input } from '@/components/ui/input';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
+import { Select, SelectContent, SelectItem, SelectTrigger } from '@/components/ui/select';
 import { Separator } from '@/components/ui/separator';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
-import { FragmentType, graphql, useFragment } from '@/gql';
-import { OperationsStats, ProjectType } from '@/gql/graphql';
+import { TooltipProvider } from '@/components/ui/tooltip';
+import { graphql } from '@/gql';
+import { ProjectType } from '@/gql/graphql';
 import { subDays } from '@/lib/date-time';
-import { toDecimal, useFormattedNumber, useFormattedThroughput } from '@/lib/hooks';
-import { cn, hexToRgba, pluralize, useChartStyles } from '@/lib/utils';
 import { UTCDate } from '@date-fns/utc';
 import { Link, useRouter } from '@tanstack/react-router';
 
@@ -50,83 +29,11 @@ export const OrganizationIndexRouteSearch = z.object({
 
 type RouteSearchProps = z.infer<typeof OrganizationIndexRouteSearch>;
 
-const ProjectCard_ProjectFragment = graphql(`
-  fragment ProjectCard_ProjectFragment on Project {
-    id
-    slug
-    type
-  }
-`);
-
-const TargetCard_TargetFragment = graphql(`
-  fragment TargetCard_TargetFragment on Target {
-    id
-    slug
-  }
-`);
-
-export const OverTimeStats_OrganizationProjectsPageFragment = graphql(`
-  fragment OverTimeStats_OrganizationProjectsPageFragment on OperationsStats {
-    failuresOverTime(resolution: $chartResolution) {
-      date
-      value
-    }
-    requestsOverTime(resolution: $chartResolution) {
-      date
-      value
-    }
-  }
-`);
-
-const TargetCardQuery = graphql(`
-  query TargetCardQuery(
-    $organizationSlug: String!
-    $projectSlug: String!
-    $targetSlug: String!
-    $chartResolution: Int!
-    $period: DateRangeInput!
-  ) {
-    target: target(
-      reference: {
-        bySelector: {
-          organizationSlug: $organizationSlug
-          projectSlug: $projectSlug
-          targetSlug: $targetSlug
-        }
-      }
-    ) {
-      id
-      slug
-      ...TargetCard_TargetFragment
-      totalRequests(period: $period)
-      requestsOverTime(resolution: $chartResolution, period: $period) {
-        date
-        value
-      }
-      schemaVersionsCount(period: $period)
-      operationsStats(period: $period) {
-        ... on OperationsStats {
-          totalRequests
-          totalFailures
-          totalOperations
-          duration {
-            p75
-            p90
-            p95
-            p99
-          }
-        }
-        ...OverTimeStats_OrganizationProjectsPageFragment
-      }
-    }
-  }
-`);
-
 const OrganizationProjectsPageQuery = graphql(`
   query OrganizationProjectsPageQuery(
     $organizationSlug: String!
-    $chartResolution: Int!
     $period: DateRangeInput!
+    $chartResolution: Int!
   ) {
     organization: organizationBySlug(organizationSlug: $organizationSlug) {
       id
@@ -136,7 +43,7 @@ const OrganizationProjectsPageQuery = graphql(`
           node {
             id
             slug
-            ...ProjectCard_ProjectFragment
+            type
             totalRequests(period: $period)
             requestsOverTime(resolution: $chartResolution, period: $period) {
               date
@@ -149,6 +56,7 @@ const OrganizationProjectsPageQuery = graphql(`
                   id
                   slug
                   totalRequests(period: $period)
+                  schemaVersionsCount(period: $period)
                 }
               }
             }
@@ -165,350 +73,27 @@ const projectTypeFullNames = {
   [ProjectType.Single]: 'Monolithic Schema',
 };
 
-const NewProjectCardTargetCardCount = (props: {
-  title: string;
-  count: string;
-  description: string;
-}) => {
-  return (
-    <div className="bg-neutral-2 dark:bg-neutral-2 border-neutral-5 dark:border-neutral-4 flex flex-col rounded-lg border p-4">
-      <div className="mb-2 text-xs font-medium">{props.title}</div>
-      <div className="text-2xl font-bold">{props.count}</div>
-      <div className="whitespace-nowrap text-xs">{props.description}</div>
-    </div>
-  );
-};
-
-const useIsInView = (ref: React.RefObject<HTMLDivElement>) => {
-  const [isInView, setIsInView] = useState(false);
-
-  useEffect(() => {
-    if (ref.current) {
-      const observer = new IntersectionObserver(([entry]) => {
-        if (entry.isIntersecting) {
-          setIsInView(entry.isIntersecting);
-        }
-      });
-
-      observer.observe(ref.current);
-
-      return () => observer.disconnect();
-    }
-  }, [ref]);
-
-  return isInView;
-};
-
-export const NewProjectCardTargetCard = (props: {
-  target: FragmentType<typeof TargetCard_TargetFragment> | null;
-  className?: string;
+const ProjectCard = (props: {
+  id: string;
+  slug: string;
+  type: ProjectType;
   highestNumberOfRequests: number;
-  requestsOverTime: { date: string; value: number }[] | null;
+  totalRequests: number | null;
   schemaVersionsCount: number | null;
   days: number;
-  organizationSlug: string;
-  projectSlug: string;
-}): ReactElement => {
-  const ref = useRef<HTMLDivElement>(null);
-  const isInView = useIsInView(ref);
-
-  const { highestNumberOfRequests } = props;
-
-  const period = {
-    from: formatISO(startOfDay(subDays(new Date(), props.days))),
-    to: formatISO(endOfDay(new Date())),
-  };
-
-  const [query] = useQuery({
-    query: TargetCardQuery,
-    variables: {
-      organizationSlug: props.organizationSlug,
-      projectSlug: props.projectSlug,
-      targetSlug: props.target?.slug ?? '',
-      chartResolution: props.days,
-      period,
-    },
-    requestPolicy: 'network-only',
-    pause: !isInView || !props.target?.slug,
-  });
-
-  const target = query.data?.target;
-
-  const requests = useMemo(() => {
-    if (target?.requestsOverTime?.length) {
-      return target.requestsOverTime.map<[string, number]>(node => [node.date, node.value]);
-    }
-
-    return [
-      [new Date(subDays(new Date(), props.days)).toISOString(), 0],
-      [new Date().toISOString(), 0],
-    ] as [string, number][];
-  }, [target?.requestsOverTime]);
-
-  const failures = useMemo(() => {
-    if (target?.operationsStats?.failuresOverTime?.length) {
-      return target.operationsStats.failuresOverTime.map<[string, number]>(node => [
-        node.date,
-        node.value,
-      ]);
-    }
-
-    return [
-      [new Date(subDays(new Date(), props.days)).toISOString(), 0],
-      [new Date().toISOString(), 0],
-    ] as [string, number][];
-  }, [target?.operationsStats?.failuresOverTime]);
-
-  const totalNumberOfRequests = useMemo(
-    () => requests.reduce((acc, [_, value]) => acc + value, 0),
-    [requests],
-  );
-  const totalNumberOfVersions = props.schemaVersionsCount ?? 0;
-  const requestsInDateRange = useFormattedNumber(totalNumberOfRequests);
-  const schemaVersionsInDateRange = useFormattedNumber(totalNumberOfVersions);
-  const rpm = useFormattedThroughput({
-    requests: totalNumberOfRequests,
-    window: differenceInMilliseconds(new Date(period.to), new Date(period.from)),
-  });
-  const totalFailures = target?.operationsStats?.totalFailures ?? 0;
-
-  const successRate = useMemo(() => {
-    return totalNumberOfRequests || totalFailures
-      ? `${toDecimal(((totalNumberOfRequests - totalFailures) * 100) / totalNumberOfRequests)}%`
-      : '-';
-  }, [totalNumberOfRequests, totalFailures]);
-
-  const failureRate = useMemo(() => {
-    return totalNumberOfRequests || totalFailures
-      ? `${toDecimal((totalFailures * 100) / totalNumberOfRequests)}%`
-      : '-';
-  }, [totalNumberOfRequests, totalFailures]);
-
-  const { colors } = useChartStyles();
-
-  const [selectedLatencyKind, setSelectedLatencyKind] = useState<'p75' | 'p90' | 'p95' | 'p99'>(
-    'p95',
-  );
-
-  return (
-    <TooltipProvider>
-      <div ref={ref} className={cn('space-y-4 p-4', props.className)}>
-        <Link
-          to="/$organizationSlug/$projectSlug/$targetSlug"
-          disabled={
-            props.organizationSlug == null || props.projectSlug == null || target?.slug == null
-          }
-          params={{
-            organizationSlug: props.organizationSlug ?? 'unknown-yet',
-            projectSlug: props.projectSlug ?? 'unknown-yet',
-            targetSlug: target?.slug ?? 'unknown-yet',
-          }}
-          className="group grid w-full grid-cols-[1fr_auto] items-center gap-8 overflow-hidden"
-        >
-          <h4 className="line-clamp-1 text-lg font-medium">{target?.slug}</h4>
-          <div className="text-neutral-11 flex items-center gap-1 group-hover:underline">
-            Open <ArrowUpRight className="size-4" />
-          </div>
-        </Link>
-        <div className="space-y-4">
-          <div className="grid grid-cols-2 gap-2">
-            <div className="text-neutral-11">Requests</div>
-            <div className="text-right font-medium">
-              {requestsInDateRange} ({rpm} RPM)
-            </div>
-          </div>
-          <div className="bg-neutral-4 dark:bg-neutral-5 h-px w-full" />
-          <div className="space-y-2">
-            <div className="grid grid-cols-2 gap-2">
-              <div className="text-neutral-11">Success rate</div>
-              <div className="text-right font-medium">{successRate}</div>
-            </div>
-            <div className="bg-neutral-2 relative h-1 w-full overflow-hidden rounded-lg">
-              <div
-                className="absolute bottom-0 right-0 top-0 h-full bg-red-900"
-                style={{
-                  width: failureRate,
-                  left: successRate,
-                }}
-              />
-              <div
-                className="absolute bottom-0 left-0 top-0 h-full bg-green-500"
-                style={{ width: successRate }}
-              />
-            </div>
-          </div>
-          <div className="bg-neutral-4 dark:bg-neutral-5 h-px w-full" />
-          <div className="grid grid-cols-2 gap-2">
-            <div className="text-neutral-11 -my-2 flex items-center gap-2">
-              Latency
-              <Select
-                value={selectedLatencyKind}
-                onValueChange={value => setSelectedLatencyKind(value as typeof selectedLatencyKind)}
-              >
-                <SelectTrigger className="w-16! h-6! px-2!">
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="p75">p75</SelectItem>
-                  <SelectItem value="p90">p90</SelectItem>
-                  <SelectItem value="p95">p95</SelectItem>
-                  <SelectItem value="p99">p99</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="text-right font-medium uppercase">
-              {target?.operationsStats?.duration?.[selectedLatencyKind]}ms
-            </div>
-          </div>
-          <div className="bg-neutral-4 dark:bg-neutral-5 h-px w-full" />
-          <div className="grid grid-cols-2 gap-2">
-            <div className="text-neutral-11 uppercase">Commits</div>
-            <div className="text-right font-medium uppercase">
-              {schemaVersionsInDateRange}{' '}
-              <span className="text-neutral-10">in the last {props.days} days</span>
-            </div>
-          </div>
-          <div className="h-24">
-            <AutoSizer>
-              {size => (
-                <ReactECharts
-                  style={{ width: size.width, height: size.height }}
-                  option={{
-                    animation: !!target,
-                    color: ['#f4b740'],
-                    grid: {
-                      left: 0,
-                      top: 10,
-                      right: 0,
-                      bottom: 10,
-                    },
-                    tooltip: {
-                      trigger: 'axis',
-                      axisPointer: {
-                        label: {
-                          formatter({ value }: { value: number }) {
-                            return new Date(value).toDateString();
-                          },
-                        },
-                      },
-                    },
-                    xAxis: [
-                      {
-                        show: false,
-                        type: 'time',
-                        boundaryGap: false,
-                      },
-                    ],
-                    yAxis: [
-                      {
-                        show: false,
-                        type: 'value',
-                        min: 0,
-                        max: highestNumberOfRequests,
-                      },
-                    ],
-                    series: [
-                      {
-                        name: 'Failures',
-                        type: 'line',
-                        smooth: false,
-                        lineStyle: {
-                          width: 2,
-                        },
-                        showSymbol: false,
-                        color: colors.error,
-                        areaStyle: {
-                          opacity: 0.8,
-                          color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-                            {
-                              offset: 0,
-                              color: colors.error,
-                            },
-                            {
-                              offset: 1,
-                              color: hexToRgba(colors.error, 0),
-                            },
-                          ]),
-                        },
-                        emphasis: {
-                          focus: 'series',
-                        },
-                        data: failures,
-                      },
-                      {
-                        name: 'Requests',
-                        type: 'line',
-                        smooth: false,
-                        lineStyle: {
-                          width: 2,
-                        },
-                        showSymbol: false,
-                        color: colors.primary,
-                        areaStyle: {
-                          opacity: 0.8,
-                          color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-                            {
-                              offset: 0,
-                              color: colors.primary,
-                            },
-                            {
-                              offset: 1,
-                              color: hexToRgba(colors.primary, 0),
-                            },
-                          ]),
-                        },
-                        emphasis: {
-                          focus: 'series',
-                        },
-                        data: requests,
-                      },
-                    ],
-                  }}
-                />
-              )}
-            </AutoSizer>
-          </div>
-        </div>
-      </div>
-    </TooltipProvider>
-  );
-};
-
-const NewProjectCard = (props: {
-  project: FragmentType<typeof ProjectCard_ProjectFragment> | null;
-  cleanOrganizationId: string | null;
-  highestNumberOfRequests: number;
-  requestsOverTime: { date: string; value: number }[] | null;
-  schemaVersionsCount: number | null;
-  days: number;
-  targets: FragmentType<typeof TargetCard_TargetFragment>[] | null;
+  targets:
+    | {
+        id: string;
+        slug: string;
+        totalRequests: number;
+        schemaVersionsCount: number;
+      }[]
+    | null;
   organizationSlug: string;
   projectSlug: string;
   sortKey: string;
   sortOrder: number;
 }) => {
-  const project = useFragment(ProjectCard_ProjectFragment, props.project);
-
-  const requests = useMemo(() => {
-    if (props.requestsOverTime?.length) {
-      return props.requestsOverTime.map<[string, number]>(node => [node.date, node.value]);
-    }
-
-    return [
-      [new Date(subDays(new Date(), props.days)).toISOString(), 0],
-      [new Date().toISOString(), 0],
-    ] as [string, number][];
-  }, [props.requestsOverTime]);
-
-  const totalNumberOfRequests = useMemo(
-    () => requests.reduce((acc, [_, value]) => acc + value, 0),
-    [requests],
-  );
-  const totalNumberOfVersions = props.schemaVersionsCount ?? 0;
-
-  const requestsInDateRange = useFormattedNumber(totalNumberOfRequests);
-  const schemaVersionsInDateRange = useFormattedNumber(totalNumberOfVersions);
-
   const sortedTargets = useMemo(() => {
     return props.targets?.sort((a, b) => {
       const diffRequests = b.totalRequests - a.totalRequests;
@@ -531,10 +116,6 @@ const NewProjectCard = (props: {
     });
   }, [props.targets, props.sortKey, props.sortOrder]);
 
-  if (!project) {
-    return null;
-  }
-
   return (
     <TooltipProvider>
       <div className="border-neutral-5 dark:border-neutral-4 bg-neutral-1 dark:bg-neutral-3 overflow-hidden rounded-lg border">
@@ -544,27 +125,49 @@ const NewProjectCard = (props: {
           className="group grid w-full grid-cols-[1fr_auto] items-center gap-8 overflow-hidden p-4"
         >
           <div className="flex gap-x-2">
-            <h4 className="line-clamp-2 text-lg font-medium">{project.slug}</h4>
-            <Badge variant="secondary">{projectTypeFullNames[project.type]}</Badge>
+            <h4 className="line-clamp-2 text-lg font-medium">{props.slug}</h4>
+            <Badge variant="secondary">{projectTypeFullNames[props.type]}</Badge>
           </div>
           <div className="text-neutral-11 flex items-center gap-1 group-hover:underline">
             Open <ArrowUpRight className="size-4" />
           </div>
         </Link>
-        <div className="border-t-neutral-6 dark:border-t-neutral-5 divide-neutral-4 dark:divide-neutral-5 -mb-1 -mr-1 grid grid-cols-[repeat(auto-fit,minmax(calc(var(--spacing)*128),1fr))] divide-x divide-y border-t">
+        <div className="border-t-neutral-5 dark:border-t-neutral-4 divide-neutral-4 dark:divide-neutral-5 -mb-1 -mr-1 grid grid-cols-[repeat(auto-fit,minmax(calc(var(--spacing)*128),1fr))] divide-x divide-y border-t">
           {sortedTargets?.map(target => (
-            <NewProjectCardTargetCard
+            <TargetCard
               key={target.id}
-              target={target}
-              highestNumberOfRequests={props.highestNumberOfRequests}
-              days={props.days}
+              id={target.id}
+              slug={target.slug}
               organizationSlug={props.organizationSlug}
               projectSlug={props.projectSlug}
+              highestNumberOfRequests={props.highestNumberOfRequests}
+              days={props.days}
             />
           ))}
         </div>
       </div>
     </TooltipProvider>
+  );
+};
+
+const ProjectCardSkeleton = () => {
+  return (
+    <div className="border-neutral-5 dark:border-neutral-4 bg-neutral-1 dark:bg-neutral-3 overflow-hidden rounded-lg border">
+      <div className="group grid w-full grid-cols-[1fr_auto] items-center gap-8 overflow-hidden p-4">
+        <div className="flex gap-x-2">
+          <div className="bg-neutral-2 dark:bg-neutral-5 h-4 w-32 rounded-lg" />
+          <Badge variant="secondary" className="w-16" />
+        </div>
+        <div className="text-neutral-11 flex items-center gap-1 group-hover:underline">
+          Open <ArrowUpRight className="size-4" />
+        </div>
+      </div>
+      <div className="border-t-neutral-6 dark:border-t-neutral-5 divide-neutral-4 dark:divide-neutral-5 -mb-1 -mr-1 grid grid-cols-[repeat(auto-fit,minmax(calc(var(--spacing)*128),1fr))] divide-x divide-y border-t">
+        {Array.from({ length: 3 }).map((_, index) => (
+          <TargetCardSkeleton key={index} />
+        ))}
+      </div>
+    </div>
   );
 };
 
@@ -782,13 +385,14 @@ function OrganizationPageContent(
             ) : (
               <div className="flex w-full flex-col gap-y-8">
                 {projects.map(project => (
-                  <NewProjectCard
+                  <ProjectCard
                     key={project.id}
-                    cleanOrganizationId={currentOrganization.slug}
+                    id={project.id}
+                    slug={project.slug}
+                    type={project.type}
                     days={days}
                     highestNumberOfRequests={highestNumberOfRequests}
-                    project={project}
-                    requestsOverTime={project.requestsOverTime}
+                    totalRequests={project.totalRequests}
                     schemaVersionsCount={project.schemaVersionsCount}
                     targets={project.targets.edges.map(edge => edge.node)}
                     organizationSlug={props.organizationSlug}
@@ -802,16 +406,7 @@ function OrganizationPageContent(
           ) : (
             <div className="flex w-full flex-col gap-y-8">
               {Array.from({ length: 4 }).map((_, index) => (
-                <NewProjectCard
-                  key={index}
-                  days={days}
-                  highestNumberOfRequests={highestNumberOfRequests}
-                  project={null}
-                  cleanOrganizationId={null}
-                  requestsOverTime={null}
-                  schemaVersionsCount={null}
-                  targets={[]}
-                />
+                <ProjectCardSkeleton key={index} />
               ))}
             </div>
           )}
