+++ packages/web/app/src/pages/project.tsx
@@ -1,12 +1,10 @@
 import { ChangeEvent, ReactElement, useCallback, useMemo, useRef } from 'react';
 import { endOfDay, formatISO, startOfDay } from 'date-fns';
-import * as echarts from 'echarts';
-import ReactECharts from 'echarts-for-react';
-import { Globe, History, MoveDownIcon, MoveUpIcon, SearchIcon } from 'lucide-react';
-import AutoSizer from 'react-virtualized-auto-sizer';
+import { MoveDownIcon, MoveUpIcon, SearchIcon } from 'lucide-react';
 import { useQuery } from 'urql';
 import { z } from 'zod';
 import { Page, ProjectLayout } from '@/components/layouts/project';
+import { TargetCard, TargetCardSkeleton } from '@/components/organization/TargetCard';
 import { Button } from '@/components/ui/button';
 import { EmptyList } from '@/components/ui/empty-list';
 import { Input } from '@/components/ui/input';
@@ -15,199 +13,11 @@ import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { Select, SelectContent, SelectItem, SelectTrigger } from '@/components/ui/select';
 import { Separator } from '@/components/ui/separator';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
-import { Card } from '@/components/v2/card';
-import { FragmentType, graphql, useFragment } from '@/gql';
+import { graphql } from '@/gql';
 import { subDays } from '@/lib/date-time';
-import { useFormattedNumber } from '@/lib/hooks';
-import { cn, pluralize } from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import { UTCDate } from '@date-fns/utc';
-import { Link, useRouter } from '@tanstack/react-router';
-import { NewProjectCardTargetCard } from './organization';
-
-const TargetCard_TargetFragment = graphql(`
-  fragment TargetCard_TargetFragment on Target {
-    id
-    slug
-  }
-`);
-
-const TargetCard = (props: {
-  target: FragmentType<typeof TargetCard_TargetFragment> | null;
-  highestNumberOfRequests: number;
-  requestsOverTime: { date: string; value: number }[] | null;
-  schemaVersionsCount: number | null;
-  days: number;
-  organizationSlug: string;
-  projectSlug: string;
-}): ReactElement => {
-  const target = useFragment(TargetCard_TargetFragment, props.target);
-  const { highestNumberOfRequests } = props;
-  const requests = useMemo(() => {
-    if (props.requestsOverTime?.length) {
-      return props.requestsOverTime.map<[string, number]>(node => [node.date, node.value]);
-    }
-
-    return [
-      [new Date(subDays(new Date(), props.days)).toISOString(), 0],
-      [new Date().toISOString(), 0],
-    ] as [string, number][];
-  }, [props.requestsOverTime]);
-
-  const totalNumberOfRequests = useMemo(
-    () => requests.reduce((acc, [_, value]) => acc + value, 0),
-    [requests],
-  );
-  const totalNumberOfVersions = props.schemaVersionsCount ?? 0;
-  const requestsInDateRange = useFormattedNumber(totalNumberOfRequests);
-  const schemaVersionsInDateRange = useFormattedNumber(totalNumberOfVersions);
-
-  return (
-    <Card
-      asChild
-      className="hover:bg-neutral-5/40 hover:shadow-neutral-5/50 bg-neutral-2/50 h-full self-start px-0 pt-4 hover:shadow-md"
-    >
-      <Link
-        to="/$organizationSlug/$projectSlug/$targetSlug"
-        disabled={
-          props.organizationSlug == null || props.projectSlug == null || target?.slug == null
-        }
-        params={{
-          organizationSlug: props.organizationSlug ?? 'unknown-yet',
-          projectSlug: props.projectSlug ?? 'unknown-yet',
-          targetSlug: target?.slug ?? 'unknown-yet',
-        }}
-      >
-        <TooltipProvider>
-          <div className="flex items-start gap-x-2">
-            <div className="grow">
-              <div>
-                <AutoSizer disableHeight>
-                  {size => (
-                    <ReactECharts
-                      style={{ width: size.width, height: 90 }}
-                      option={{
-                        animation: !!target,
-                        color: ['#f4b740'],
-                        grid: {
-                          left: 0,
-                          top: 10,
-                          right: 0,
-                          bottom: 10,
-                        },
-                        tooltip: {
-                          trigger: 'axis',
-                          axisPointer: {
-                            label: {
-                              formatter({ value }: { value: number }) {
-                                return new Date(value).toDateString();
-                              },
-                            },
-                          },
-                        },
-                        xAxis: [
-                          {
-                            show: false,
-                            type: 'time',
-                            boundaryGap: false,
-                          },
-                        ],
-                        yAxis: [
-                          {
-                            show: false,
-                            type: 'value',
-                            min: 0,
-                            max: highestNumberOfRequests,
-                          },
-                        ],
-                        series: [
-                          {
-                            name: 'Requests',
-                            type: 'line',
-                            smooth: false,
-                            lineStyle: {
-                              width: 2,
-                            },
-                            showSymbol: false,
-                            areaStyle: {
-                              opacity: 0.8,
-                              color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
-                                {
-                                  offset: 0,
-                                  color: 'rgba(244, 184, 64, 0.20)',
-                                },
-                                {
-                                  offset: 1,
-                                  color: 'rgba(244, 184, 64, 0)',
-                                },
-                              ]),
-                            },
-                            emphasis: {
-                              focus: 'series',
-                            },
-                            data: requests,
-                          },
-                        ],
-                      }}
-                    />
-                  )}
-                </AutoSizer>
-              </div>
-              <div className="flex flex-row items-center justify-between gap-y-3 px-4 pt-4">
-                <div>
-                  {target ? (
-                    <h4 className="line-clamp-2 text-lg font-bold">{target.slug}</h4>
-                  ) : (
-                    <div className="bg-neutral-5 h-4 w-48 animate-pulse rounded-full py-2" />
-                  )}
-                </div>
-                <div className="flex flex-col gap-y-2 py-1">
-                  {target ? (
-                    <>
-                      <Tooltip>
-                        <TooltipTrigger>
-                          <div className="flex flex-row items-center gap-x-2">
-                            <Globe className="text-neutral-10 size-4" />
-                            <div className="text-xs">
-                              {requestsInDateRange}{' '}
-                              {pluralize(totalNumberOfRequests, 'request', 'requests')}
-                            </div>
-                          </div>
-                        </TooltipTrigger>
-                        <TooltipContent>
-                          Number of GraphQL requests in the last {props.days} days.
-                        </TooltipContent>
-                      </Tooltip>
-                      <Tooltip>
-                        <TooltipTrigger>
-                          <div className="flex flex-row items-center gap-x-2">
-                            <History className="text-neutral-10 size-4" />
-                            <div className="text-xs">
-                              {schemaVersionsInDateRange}{' '}
-                              {pluralize(totalNumberOfVersions, 'commit', 'commits')}
-                            </div>
-                          </div>
-                        </TooltipTrigger>
-                        <TooltipContent>
-                          Number of schemas pushed to this project in the last {props.days} days.
-                        </TooltipContent>
-                      </Tooltip>
-                    </>
-                  ) : (
-                    <>
-                      <div className="bg-neutral-5 my-1 h-2 w-16 animate-pulse rounded-full" />
-                      <div className="bg-neutral-5 my-1 h-2 w-16 animate-pulse rounded-full" />
-                    </>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-        </TooltipProvider>
-      </Link>
-    </Card>
-  );
-};
+import { useRouter } from '@tanstack/react-router';
 
 export const ProjectIndexRouteSearch = z.object({
   search: z.string().optional(),
@@ -432,15 +242,14 @@ const ProjectsPageContent = (
             />
           ) : (
             targets.map(target => (
-              <NewProjectCardTargetCard
+              <TargetCard
                 key={target.id}
-                target={target}
-                days={days}
-                highestNumberOfRequests={highestNumberOfRequests}
-                requestsOverTime={target.requestsOverTime}
-                schemaVersionsCount={target.schemaVersionsCount}
+                id={target.id}
+                slug={target.slug}
                 organizationSlug={props.organizationSlug}
                 projectSlug={props.projectSlug}
+                days={days}
+                highestNumberOfRequests={highestNumberOfRequests}
                 className="bg-neutral-1 dark:bg-neutral-3 border-neutral-4 dark:border-neutral-5 rounded-lg border"
               />
             ))
@@ -448,15 +257,8 @@ const ProjectsPageContent = (
         ) : (
           <>
             {Array.from({ length: 4 }).map((_, index) => (
-              <NewProjectCardTargetCard
+              <TargetCardSkeleton
                 key={index}
-                target={null}
-                days={days}
-                highestNumberOfRequests={highestNumberOfRequests}
-                requestsOverTime={null}
-                schemaVersionsCount={null}
-                organizationSlug={props.organizationSlug}
-                projectSlug={props.projectSlug}
                 className="bg-neutral-1 dark:bg-neutral-3 border-neutral-6 dark:border-neutral-5 rounded-lg border"
               />
             ))}
@@ -479,7 +281,6 @@ const ProjectOverviewPageQuery = graphql(`
         node {
           id
           slug
-          ...TargetCard_TargetFragment
           totalRequests(period: $period)
           requestsOverTime(resolution: $chartResolution, period: $period) {
             date
